requirements.txt
streamlit>=1.21.0
altair<5
google-generativeai>=0.8.0
pytest
